ID2 fixes based on review
