+++ packages/common/src/types.ts
@@ -55,6 +55,21 @@ export enum AttributeKeyString {
     taxIdNo = 'taxIdNo',
 }
 
+export enum Sex {
+    NotKnown = '0',
+    Male = '1',
+    Female = '2',
+    NA = '9',
+}
+
+export enum IdDocType {
+    NA = '0',
+    Passport = '1',
+    NationalIdCard = '2',
+    DriversLicense = '3',
+    ImmigrationCard = '4',
+}
+
 export enum TransactionStatusEnum {
     Received = 'received',
     Finalized = 'finalized',
